Drop the browser field from the package.json
I think I made a mistake by adding this in the first place.

Having the browser field point to a CommonJS build keeps ESM-aware bundlers from going to the "module" field like we want them to.

For the non-ESM bundlers, the UMD build only adds a couple hundred bytes or so.